Implemeted paystack fully and unit assignment

+++ SMSBreeze.Web/Models/Entities/Customer.cs
@@ -16,7 +16,7 @@ namespace SMSBreeze.Models.Entities
         [DataType(DataType.Date)]
         [DisplayFormat(DataFormatString = "{0:yyyy-MM-dd}", ApplyFormatInEditMode = true)]
         public DateTime DateCreated { get; set; }
-		public decimal? SmsBalance { get; set; }
+		public int? SmsBalance { get; set; }
 		public ApplicationUser  ApplicationUser { get; set; }
 		public List<SmsTransaction> SmsTransactions { get; set; }
 		public List<Contact> Contacts { get; set; }
